post-to-bluesky/requirements.txt
python-dotenv
atproto>=0.0.55
jsonpath-ng
